fix: Properly detect taggability of remote modules
Changes in this commit make sure that Yor searches in the right folder
when determining if module in question is taggable.

For example, before this commit, Yor will skip the policy module:

    module "policy" {
      source  = "terraform-aws-modules/iam/aws//modules/iam-policy"
      version = "5.5.4"
      # more here
    }

The problem was that Yor was searching for the variables block in the
.terraform/modules/policy/*.tf, ignoring the fact that module source
tells us that we need to use modules/iam-policy subdirectory.

To remedy the situation, We added a new function that extracts
subdirectory from the source and appends it to the search path.