Add check mode for yor tag command
This commit adds a "check" mode to the tag command, which causes yor
to exit with a non-zero status (by calling log.Error()) if any changes
were (or would have been, if in dry mode) made.

This is useful in CI pipelines, where one wants to confirm that
everything is tagged appropriately.  Check mode is simpler to use in
CI than parsing the CLI output.

The check for changes is made after the report is printed, as knowing
what was/would-have-been changed is useful in debugging CI failures.

This commit:

- adds a utility method to the report service which returns a boolean
  value if there were any additions or changes to the tags;
- uses that utility method in the tag command; and
- adds a single simple test for the utility function.

Quibbles:

- It would be nice to be able to test the utility function in a
  scenario with no changes, but there does not seem to be a test
  helper to generate that scenario.
- There is no test for the CLI option handling, I'm not clear on how
  the CLI is tested.
- It might be useful to add a Quiet flag, which suppresses all output.
  This seems like an orthogonal addition, not part of this MR.